renames KerberosCredentials
