+++ httpclient5-testing/src/test/java/org/apache/hc/client5/testing/compatibility/sync/HttpClientCompatibilityTest.java
@@ -31,8 +31,8 @@ import org.apache.hc.client5.http.ContextBuilder;
 import org.apache.hc.client5.http.auth.AuthScope;
 import org.apache.hc.client5.http.auth.Credentials;
 import org.apache.hc.client5.http.auth.CredentialsStore;
-import org.apache.hc.client5.http.auth.KerberosCredentials;
 import org.apache.hc.client5.http.auth.UsernamePasswordCredentials;
+import org.apache.hc.client5.http.auth.gss.GssCredentials;
 import org.apache.hc.client5.http.classic.methods.HttpGet;
 import org.apache.hc.client5.http.classic.methods.HttpOptions;
 import org.apache.hc.client5.http.impl.auth.BasicCredentialsProvider;
@@ -68,7 +68,7 @@ public abstract class HttpClientCompatibilityTest {
         this.clientResource = new HttpClientResource();
         if (targetCreds != null) {
             //this.setCredentials(new AuthScope(target), targetCreds);
-            if (targetCreds instanceof KerberosCredentials) {
+            if (targetCreds instanceof GssCredentials) {
                 secretPath = "/private_spnego/big-secret.txt";
                 this.clientResource.configure(builder -> builder
                     .setTargetAuthenticationStrategy(new SpnegoAuthenticationStrategy())
@@ -79,7 +79,7 @@ public abstract class HttpClientCompatibilityTest {
             this.clientResource.configure(builder -> builder.setProxy(proxy));
             if (proxyCreds != null) {
                 this.setCredentials(new AuthScope(proxy), proxyCreds);
-                if (proxyCreds instanceof KerberosCredentials) {
+                if (proxyCreds instanceof GssCredentials) {
                     // We disable Mutual Auth, because Squid does not support it.
                     // There is no way to set separate scheme registry for target/proxy,
                     // but that's not a problem as SPNEGO cannot be proxied anyway.
