+++ httpclient5/src/main/java/org/apache/hc/client5/http/impl/auth/gss/GssSchemeBase.java
@@ -313,9 +313,9 @@ public abstract class GssSchemeBase implements AuthScheme {
         }
         final Credentials credentials =
                 credentialsProvider.getCredentials(new AuthScope(host, null, getName()), context);
-        if (credentials instanceof org.apache.hc.client5.http.auth.KerberosCredentials) {
+        if (credentials instanceof org.apache.hc.client5.http.auth.gss.GssCredentials) {
             this.gssCredential =
-                    ((org.apache.hc.client5.http.auth.KerberosCredentials) credentials)
+                    ((org.apache.hc.client5.http.auth.gss.GssCredentials) credentials)
                             .getGSSCredential();
         } else {
             this.gssCredential = null;
